#771 - Entropie Reductor: Polar-Polish (DTO-Float-Fix, Polar-Label, Polar-Empty-State)
Live-Sonde 2026-05-16 nach Polar-Verbindung deckte drei Punkte auf:

1. PolarUser-DTO Float-Bug: Polar's GET /v3/users liefert "height":178.0
   (Float), mein DTO hatte Int? — JSON-Parser warf "Unexpected symbol '.'
   in numeric literal". Folge: registerUser warf eine Exception, aber der
   Sync lief trotzdem durch (Defensive). Fix: weight + height beide als
   Float? typisiert. Polar's Praezision (88.2 kg, 178.0 cm) bleibt erhalten.

2. Quellenabhaengiges Label: Der hardcoded "T-Rex 3"-Chip in der Detail-
   Card und der "Training (T-Rex 3)"-Default-Titel zeigten die alte
   Quelle auch fuer Polar-Workouts. Neue sourceLabel()-Helper-Funktion:
   - source="polar"       -> "Polar"
   - source="health_connect" -> "Health Connect"
   - source enthaelt "huami" -> "T-Rex 3" (Zepp-Cloud)
   - null/leer            -> "T-Rex 3" (Legacy-Default)
   - sonst                -> "Polar" (defensiver Default fuer Zukunft)

3. Empty-State im Trainings-Tab: "Verbinde dein Zepp-Konto" zeigte den
   alten Hinweis auch fuer Polar-only-User. Umgestellt auf "Verbinde
   dein Polar-Konto" mit Hinweis auf H10-Brustgurt-Daten.

Verifikation Logcat (2026-05-16 14:01:57):
- Polar-Token-Exchange OK, x_user_id=10211799 gespeichert
- POST /v3/users -> 201 Created
- POST /exercise-transactions -> 204 No Content (kein Backlog)
- PolarSyncWorker SUCCESS (kein Retry)
Frank's Polar-Flow-Konto hat aktuell keine neuen Trainings zum Abholen.
Sobald er mit seiner Polar-Uhr (Pacer Pro/Vantage M/V3/Ignite 3) ein
Training aufzeichnet, kommt es beim naechsten Sync runter.

+++ EntropieReductor/app/src/main/java/de/frank/entropyreducer/data/remote/polar/PolarApi.kt
@@ -164,8 +164,12 @@ data class PolarUser(
     @SerialName("last-name") val lastName: String? = null,
     val birthdate: String? = null,
     val gender: String? = null,
+    /** Gewicht in kg. Live-Sonde 2026-05-16: Polar liefert Floats wie 88.2 */
     val weight: Float? = null,
-    val height: Int? = null,
+    /** Groesse in cm. Live-Sonde 2026-05-16: Polar liefert Floats wie 178.0 —
+     *  NICHT Int wie ich initial annahm. JSON-Parser warf "Unexpected symbol '.'
+     *  in numeric literal" weil 178.0 nicht in Int passt. Float ist sicher. */
+    val height: Float? = null,
 )
 
 /** Antwort auf POST /exercise-transactions — enthaelt die Transaction-ID. */
